requirements.txt
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.1
plotly==5.7.0
scikit_learn==1.1.1
streamlit==1.8.1
